Refactor game scene to use MVVM.
